Add files via upload
Added support for using the exported settings for mantuik06 as the values for the tone mapping operators. simply place the txt you generate by exporting setting in Luminance hdr in the same folder as the base t2vert script (where the .MLV files your converting are located) and it will pull the data from that .txt file as the values used by luminance hdr when tone mapping.   could cause a crash if no .txt file is present.  As of this writing only the mantiuk06 tone mapping orerator is supported, any other operators will behave strangely